New RPC ActiveIndices for the coligate to use.
The RPC returns information about the active indices of
segment reservations that have this AS as a source.
Unify queries referring to EER.
Add a new method in the DB to retrieve active EERs in the source AS.
Add DB test.
Fix: bug in colibri test functions restoring the token.